jquery slide in entries, modify composite view attachsubview

+++ db/migrate/20150208010630_create_ratings.rb
@@ -1,9 +1,9 @@
 class CreateRatings < ActiveRecord::Migration
   def change
     create_table :ratings do |t|
-      t.integer :entry_val, null: false
+      t.float :entry_val, null: false
       t.integer :entry_id, null: false
-      
+
       t.timestamps
     end
   end
